Make events callback based instead of constant polling

+++ src/backend/events.rs
@@ -16,7 +16,7 @@ pub enum WindowEvents {
     /// Will not be emitted on Wayland, mobile and web!
     WindowPos(Position<i32>),
     WindowSize(Dimensions),
-    CursorPos(Position<u32>),
+    CursorPos(Position<f64>),
     // Button number, KeyAction
     MouseButton(u32, KeyAction),
     /// x offset, y offset
